requirements.txt
SQLAlchemy
Flask
requests
WTForms
pillow
Werkzeug
Flask-WTF
flask-login
SQLAlchemy-serializer